🚚 move: 각 Page -> index.tsx
