--- app/src/pages/FtOAuthPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FtOAuthPage';
