--- app/src/pages/LandingPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LandingPage';
