--- app/src/pages/LeaderBoardPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LeaderBoardPage';
