--- app/src/pages/LogoutPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LogoutPage';
