--- app/src/pages/ProfileGeneralTab/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProfileGeneralTab';
